Added ability to pick up and drop items of a certain tag (Collectable)
If an item is given the tag of Collectable within it's BP now it can be held in the players hand.

BUGS: Item does not appear in players hand properly yet. When item is dropped it will likely get stuck in the floor, but still able to be picked up. If the player stand atop a recently dropped item and attempts to pick it back up, the player will be teleported several feet away and in the air.